refactor: 삭제될 파일 명시 및 리뷰 반영
