+++ Dear/Dear/ViewController.swift
@@ -12,7 +12,5 @@ class ViewController: UIViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
     }
-
-
 }
